+++ src/nvlib/gui/tree_window/tree_viewer.py
@@ -254,6 +254,9 @@ class TreeViewer(ttk.Frame, Observer, TreeViewerCtrl):
         self.tree.bind(KEYS.DELETE[0], self._ctrl.delete_elements)
         self.tree.bind(MOUSE.RIGHT_CLICK, self.open_context_menu)
         self.tree.bind(MOUSE.MOVE_NODE, self._on_move_node)
+        self.tree.bind(KEYS.CUT[0], self._ctrl.cut_element)
+        self.tree.bind(KEYS.COPY[0], self._ctrl.copy_element)
+        self.tree.bind(KEYS.PASTE[0], self._ctrl.paste_element)
 
     def _browse_tree(self, node):
         """Select and show node.